Modularise files
