+++ drivers/files/create.js
@@ -1,0 +1,17 @@
+// create: async function (type, id, fileContent) {
+//   // try {
+//   //   if (!type || !fileContent || !id) throw new Error('markdown.create: Missing parameters')
+//   //   if (is.not.string(type) || is.not.string(id)) throw new Error('markdown.create: Parameters must be supplied as strings')
+
+//   //   let destination = path.join(config.contentRoot(), type, id, 'index.md')
+
+//   //   fs.stat(destination, (err, stat) => {
+//   //     if (is.not.falsy(err)) throw new Error('markdown.create: File already exists')
+//   //   });
+
+//   //   return await fs.promises.writeFile(destination, message, 'utf8')
+//   // } catch (error) {
+//   //   // TODO Add to error log
+//   //   return Promise.reject(error)
+//   // }
+// },
