+++ drivers/files/delete.js
@@ -1,0 +1,15 @@
+// delete: async function (type, id) {
+//   // try {
+//   //   if (!type || !id)
+//   //     throw new Error('markdown.del: Missing parameters')
+//   //   if (is.not.string(type) || is.not.string(id))
+//   //     throw new Error('markdown.del: Parameters must be supplied as strings')
+
+//   //   let destination = path.join(config.contentRoot(), location, id, 'index.md')
+
+//   //   return await fs.promises.unlink(destination)
+//   // } catch (error) {
+//   //   return Promise.reject(error)
+//   // }
+// }
+// //
