+++ drivers/files/index.js
@@ -1,0 +1,3 @@
+const read = require('./read')
+
+module.exports = {read}
