+++ drivers/files/update.js
@@ -1,0 +1,16 @@
+// update: async function (type, id, fileContent) {
+//   // try {
+//   //   if (!type || !fileContent || !id) throw new Error('markdown.update: Missing parameters')
+//   //   if (is.not.string(type) || is.not.string(id)) throw new Error('markdown.update: Parameters must be supplied as strings')
+
+//   //   let destination = path.join(config.contentRoot(), type, id, 'index.md')
+
+//   //   fs.stat(destination, (err, stat) => {
+//   //     if (is.falsy(err)) throw new Error('markdown.update: No file exists to update')
+//   //   });
+
+//   //   return await fs.promises.writeFile(destination, message, 'utf8')
+//   // } catch (error) {
+//   //   return Promise.reject(error)
+//   // }
+// },
